Expand OAI tests to check actual returned result
A successful OAI-PMH call always returns a 200 response containing a
payload with the OAI-PMH root tag. This is regardless of the query
returning an actual result. This means that for example a request with
an "unknown verb" will also return a 200. The initial tests just checked
for that root tag and the OK status without further examining the payload.

The tests have been expanded to check if the query actually returned the
expected results.