DEV-1742 Make MH auth configurable per environment
For every MediaHaven environment, it fetches the same credentials from
the Vault. However, every environment has its own credentials. Make it
configurable, so that the specific credentials for the enviroment are
used.

Instead of fixing some assertions in the "Test current user" test, the
assertions are removed as they are not really useful.